partially fixed circ imports
