+++ napari/_app/context/_layerlist_context.py
